renamed cal.com call to cal.com video, added og image (#1852)
