+++ apps/web/pages/video/meeting-ended/[uid].tsx
@@ -13,18 +13,19 @@ import { inferSSRProps } from "@lib/types/inferSSRProps";
 import { HeadSeo } from "@components/seo/head-seo";
 import Button from "@components/ui/Button";
 
-export default function MeetingNotStarted(props: inferSSRProps<typeof getServerSideProps>) {
+export default function MeetingUnavailable(props: inferSSRProps<typeof getServerSideProps>) {
   const router = useRouter();
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const [is24h, setIs24h] = useState(false);
 
   //if no booking redirectis to the 404 page
   const emptyBooking = props.booking === null;
   useEffect(() => {
     if (emptyBooking) {
-      router.push("/call/no-meeting-found");
+      router.push("/video/no-meeting-found");
     }
   });
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const [is24h, setIs24h] = useState(false);
   if (!emptyBooking) {
     return (
       <div>
@@ -47,7 +48,7 @@ export default function MeetingNotStarted(props: inferSSRProps<typeof getServerS
                     </div>
                     <div className="mt-3 text-center sm:mt-5">
                       <h3 className="text-lg font-medium leading-6 text-gray-900" id="modal-headline">
-                        This meeting has not started yet
+                        This meeting is in the past.
                       </h3>
                     </div>
                     <div className="mt-4 border-t border-b py-4">
@@ -59,11 +60,6 @@ export default function MeetingNotStarted(props: inferSSRProps<typeof getServerS
                         {dayjs(props.booking.startTime).format(
                           (is24h ? "H:mm" : "h:mma") + ", dddd DD MMMM YYYY"
                         )}
-                      </p>
-                    </div>
-                    <div className="mt-3 text-center sm:mt-5">
-                      <p className="text-sm text-gray-500">
-                        This meeting will be accessible 60 minutes in advance.
                       </p>
                     </div>
                   </div>
